topic_agent/requirements.txt
langgraph>=0.2.30
google-generativeai>=0.7.2
requests>=2.32.0
SQLAlchemy>=2.0.32
python-dotenv>=1.0.1
selenium>=4.18.0
webdriver-manager>=4.0.0
pandas>=2.0.0
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
ddgs>=8.1.1


